implemented builtin (len) for the compiler.
